refactor(web): move the import predicates into a leaf module
isImportRedundant and isDuplicateImport are pure functions over a Factory, but
they sat in inputs.ts, which imports factory.ts, which imports problems.ts. The
incoming status module needs both and cannot reach them without closing that
cycle, so they move to inputs-analysis.ts and are re-exported from inputs.ts.
No behaviour change; inputs.spec.ts passes untouched.
