+++ web/src/utils/factory-management/inputs.ts
@@ -4,6 +4,9 @@ import { calculateFactory, findFac } from '@/utils/factory-management/factory'
 import { recalculateFactoryDependencies } from '@/utils/factory-management/dependencies'
 import { DataInterface } from '@/interfaces/DataInterface'
 import eventBus from '@/utils/eventBus'
+// Re-exported so existing callers keep importing them from here; they live in a leaf module
+// because status.ts needs them and cannot import anything that reaches factory.ts.
+export { isDuplicateImport, isImportRedundant } from '@/utils/factory-management/inputs-analysis'
 
 export const getInput = (factory: Factory, part: string, factoryId?: number) => {
   // Returns a SINGULAR input object by the outputPart. If multiple are detected, throw.
@@ -214,88 +217,6 @@ export const calculateAbleToImport = (factory: Factory, importCandidates: Factor
   }
 
   return true
-}
-
-export const isImportRedundant = (importIndex: number, factory: Factory): boolean | null => {
-  const input = factory.inputs[importIndex]
-  if (!input?.outputPart) {
-    return null
-  }
-
-  if (input.amount === 0) {
-    return null // If the amount is 0, it's technically redundant, but it could also be the user hasn't chosen anything yet. They already get a chip saying no amount is set.
-  }
-
-  const partData = factory.parts[input.outputPart]
-
-  if (!partData) {
-    console.error(`inputs: isImportRedundant: Part data for part ${input.outputPart} not found in factory ${factory.id}!`)
-    return null
-  }
-
-  // If the factory is producing the products internally, and the amount that it produces exceeds the amount imported, then the import is redundant.
-
-  const required = partData.amountRequired
-  const produced = partData.amountSuppliedViaProduction
-
-  // The remainder of the part that needs to be imported
-  const importsNeeded = required - produced
-
-  // If there's no requirement, then the import is redundant.
-  if (required <= 0) {
-    return true
-  }
-
-  // If there is sufficient internal production, then all imports are redundant
-  if (importsNeeded <= 0) {
-    return true
-  }
-
-  // Now, we also need to take into account other imports. If other imports fully satisfy the requirement, then this import is redundant.
-  // Loop through all the inputs and see if the other imports fully satisfy the requirement.
-  const otherImports = factory.inputs.filter((_, index) =>
-    index !== importIndex &&
-    factory.inputs[index].outputPart === input.outputPart
-  )
-  const otherImportsValues: number[] = []
-  otherImports.forEach(input => {
-    if (!input.outputPart) return 0
-    otherImportsValues.push(input.amount ?? 0)
-  })
-  const otherImportsTotal = otherImportsValues.reduce((acc, val) => acc + val, 0)
-
-  // If there are no other imports then the import is required.
-  if (otherImports.length === 0) return false
-
-  // In a multi-input scenario, if there's an over supply, inform the user one of their imports are redundant.
-  // Try to be deterministic by favouring the largest import.
-  const largestOtherImport = Math.max(...otherImportsValues)
-
-  // If the current import is the largest, then it's not redundant.
-  // This does annoyingly mean that if they are both EXACTLY the same, both will be redundant. Can't really get around it.
-  if (input.amount >= largestOtherImport) return false
-
-  const requirementAfterOtherImports = importsNeeded - otherImportsTotal
-
-  // If the other imports don't fully satisfy the requirement, then the import is not redundant.
-  return requirementAfterOtherImports <= 0
-}
-
-// Two rows importing the same part from the same factory collapse into a single export
-// request, so the provider only ever sees one of them. Reports whether the row at
-// inputIndex has become such a duplicate.
-export const isDuplicateImport = (factory: Factory, inputIndex: number): boolean => {
-  const input = factory.inputs[inputIndex]
-
-  if (!input?.factoryId || !input.outputPart) {
-    return false // Still being filled in.
-  }
-
-  return factory.inputs.some((other, index) =>
-    index !== inputIndex &&
-    other.factoryId === input.factoryId &&
-    other.outputPart === input.outputPart
-  )
 }
 
 export const satisfyImport = (importIndex: number, factory: Factory): void | null => {
